fixed #77

+++ makefile
@@ -44,4 +44,4 @@ $(BUILD_PATH)/kwm_template.plist: $(KWM_PLIST)
 
 $(HOME)/.kwm/kwmrc: $(SAMPLE_CONFIG)
 	mkdir -p $(CONFIG_DIR)
-	cp -n $^ $@
+	test ! -e $@ && cp -n $^ $@
